Procedural everything
All the procedural generation of artefacts, indestructables, and obstacles are done. They have not yet been converted to Bolt code : ToDo in branch.
I have optimised the code as best as I can, as such now there is a lot less gameObjects spawned and kept in the world. One cause for concern is that there is a delay in generating a 100x100 grid. The profiler states that it is 90% unityEditor so hopefully it won't be a problem in the build but do keep an eye on it.