LODS and player obstacle destruction
I've given each obstacle and indestructable a proper LOD so they know don't cause much of an effect on the system, however the game still runs abysmally slow, SEE DISCORD IMG FOR MORE INFO>

I've made it so then the player casts a sphere instead of a line for destruction.

I've changed some prefab settings.

I've gave the player a physics material in an attempt to stop the player from sliding around however it has not worked.