Generations
Found inconsistancies in obstacle and indestructable generation. They are now fixed and the players will all see the same map as eachother. EXCEPT for the scale of the indestructables.

Artefacts are all spawning however the bug where players cannot pick up the artefacts are back. It says that it is picked up a [blank] for 0 points. The object still stays in the world and does not disappear.