Generation Fix V1.0
I've tried placing the generators in a more safe position - the grass gen was possibly clipping a spawn position between the grass biome and the temple.

I've replaced the indestructible generator from being a networked function to a mono function that every client must do themselves. To do this i've edited the proto_procedural script to allow for a seed to be implemented, this should allow for syncing across all clients.

MapGenerator: indestructable now has a random scale thanks to the new system. I've also replaced a lot of the Random creation with just one random created and passed as a parameter into other functions. This is just safer than generating new randoms each time, also more resource friendly.

Currently testing the GenerateAll script. I'm going to see if I can just call the generate function from an event for everyone including the clients and server - I felt like for the server it might've been trying to call it twice.