Version Number works
I've managed to get the version number working however, people can join any game in progress. This is due to how the connection handling is set up:

Connection handling is only handled in the menu screens. As soon as we go into a game, it stops handling it manually, because there is no script there. So instead of writing a duplicate script for in game use I will be completely overhauling the scripts.

There will be one main script that I will be making to accept, refuse any incoming connections.

So this commit is for a rollback just in case things go haywire.