Re-work of networking
NetworkConnections.cs controls the authorisation of connections and handles rejections.
It holds the info of the lobby at all times, whether it is joinable, private, the password, and roomname AND the versionNo. This must be changed every release version to avoid connecting wrong versions of the game.

There's a new networkconnections script that handles connect requests, this script doesn't get destroyed.

LobbyUIManager and BrowseScreenUI is now inheriting from network connections.
Server info now has joinableSession as a variable.