Player Animation, Camera and cleanup
The animation is now hooked for the moving and the obstacle interaction. The player now gets camera in the first 5 seconds, this is necessary as OnStartAuthority is called before we enter the scene.

I've cleaned up the variables in player.

I've had a mess around with camera, and I'm going to make a branch in the future making multiple iterations of different camera types so we can test them out.