Generation Fix Summary
The map generation is now no longer networked, and has been improved and fixed upon.

Firstly the non networked version has seen a massive boost in fps, and it is very stable in terms of syncing. Due to Mirror being more reliable, we rarely lose data about object destruction, and I have not seen any inconsistencies with that.

The seed that the map generates from now appears to be completely synced, unlike before where there were one or two odd unsynced obstacles.

The randomness of the seeds have been changed, before it would have many different but unknown amount of possibilities, but the problem was that it was inconsistant. At least with the old solution and mirror it was giving very similar maps in the same hour. Now we use milliseconds as seeds, so we have a possible 1000 combinations of maps. We can expand on this further if we wanted to by maybe adding on seconds(turned to milliseconds) which would add 60k combinations. For now this is fine.

Due to the System.Random call the load time is a slight bit slower in to the game.

The generation is now being done in sequences of indestructable first, then 3 seconds after everything else for that biome. Each generator activating in 3 second intervals. This means that when Biome 1 spawns everything, Biome 2 is spawning indestructible.

Had a look at the indestructible colliders and some of them were turned off for some reason, turned them on and changed the colliders to be more tight around the base of the object. Contemplating on changing the rock colliders to mesh colliders to hopefully avoid climbing on them.

The indestructible spawning system has now been fixed. Before it would be spawning them in a 1D line at the far end of the map, now it is properly spawning them around the entire map as such the map now looks much more exciting, diverse and full.

The issue of obstacles spawning inside the Indestructables have been found, and has been fixed. HOWEVER now the map also generates on map borders at the top of the hills. This needs to be corrected by asking Joseph to re-import the same map but cut the outside into a separate mesh than the inside ones, and hopefully also the mesh where the temple is can be added to the outside mesh object so we can easily avoid spawning obstacles inside the temple.

I found one instance of where a gem spawned on a slope too steep and where there wasn't a obstacles. Keep an eye out for that. For reference it happened near the border from the jungle into the swamp, inside the swamp biome.