Lobby and Main menu commented
I've commented everything I could and made sure it was very clear in what message it was. I've added many 'JoeComment' to signify something that i've wanted to communicate to Joe.

If it has no message, usually that means that I need Joe to comment that area because I couldn't understand it, other times it may have a message asking if something is necessary and if we could delete it.