Camera Change and few fixes
Small fixes:
- Fern hitbox #138
- Player didn't actually increase falling speed

Camera Changes:
We now have 3 different camera variations however I have put no 2 in use for the game scene.
The player can control the camera with just their movement or they can use the right mouse click to pan left and right, or the middle mouse click to pan up and down.
Furthermore, we no longer have the player rotate along the axis of the plane. This was not only unnecessary but caused more issues than helped with player experience.