Testing: UI Changes Fix
While we can connect to the lobbies now, when a second player joins the lobby the host unfortunately gets the entire ui screen reset to the default.
The client appears to be fine.

This change is simply resetting it back to when it worked before the ui changes. It's a simple test to see what happens.