File Cleanup
Just tidying up the files a little bit.
One major note is that around the assets folder there are folders names CheckIfInUse
I believe the stuff in these files are no longer used, however to be safe I would like Joe to double check it.
If all the stuff is okay to delete, just go ahead and delete it.